updated API Request Height/Width to be a decimal from short

+++ SignRequestExpressAPI/Entities/RequestEntity.cs
@@ -47,9 +47,9 @@ namespace SignRequestExpressAPI.Entities
 
         public bool IsVertical { get; set; }
 
-        public short HeightInch { get; set; }
+        public decimal HeightInch { get; set; }
 
-        public short WidthInch { get; set; }
+        public decimal WidthInch { get; set; }
 
         public Guid? TemplateFK { get; set; }
